pyproject.toml
[project]
name = "03-gemini-files-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.50.1",
    "chainlit>=1.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.52.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
